convert browser test runner to ESM
